update partner with tags, removed update partner tag, exception if delete tag has still a reference to a partner

+++ backend/business-partner-agent/src/main/java/org/hyperledger/bpa/impl/TagService.java
@@ -84,6 +84,10 @@ public class TagService {
     }
 
     public void deleteTag(@NonNull UUID id) {
+        int refs = tagRepo.countReferencesToPartner(id);
+        if(refs > 0) {
+            throw new WrongApiUsageException("Tag is still used by " + refs + " partner(s");
+        }
         tagRepo.findById(id).ifPresent(s -> tagRepo.deleteById(id));
     }
